Better support for unset attributes in SelectionType
When an ezselection is added to an existing object the data_string
will be empty (ie. not the default value). To aid in this any value
of 0 will also include these unset entries.
In addition a new UnsetValue class is available which when used as
the value of a selection filter will only returns attributes which
are unset. As a shortcut the string 'unset' can also be used.